+++ .hooks/pre-commit
@@ -1,3 +1,3 @@
 #!/bin/sh
 echo 'If you see [pycodestyle] Messages, you need to fix and commit again'
-pycodestyle ozp ozpcenter ozpiwc plugins plugins_util tests --ignore=E501,E123,E128,E121,E124,E711,E402,E722 --show-source
+pycodestyle ozp ozpcenter ozpiwc plugins tests --ignore=E501,E123,E128,E121,E124,E711,E402,E722 --show-source
