+++ ozpcenter/api/listing/serializers.py
@@ -12,9 +12,9 @@ from rest_framework import serializers
 from ozpcenter import constants
 from ozpcenter import models
 from ozpcenter import errors
-from plugins_util import plugin_manager
-from plugins_util.plugin_manager import system_has_access_control
-from plugins_util.plugin_manager import system_anonymize_identifiable_data
+from plugins import plugin_manager
+from plugins.plugin_manager import system_has_access_control
+from plugins.plugin_manager import system_anonymize_identifiable_data
 import ozpcenter.api.agency.model_access as agency_model_access
 import ozpcenter.api.category.model_access as category_model_access
 import ozpcenter.api.contact_type.model_access as contact_type_model_access
