+++ ozpcenter/api/notification/serializers.py
@@ -7,8 +7,8 @@ from django.core.exceptions import ObjectDoesNotExist
 from rest_framework import serializers
 
 from ozpcenter import models
-from plugins_util import plugin_manager
-from plugins_util.plugin_manager import system_anonymize_identifiable_data
+from plugins import plugin_manager
+from plugins.plugin_manager import system_anonymize_identifiable_data
 import ozpcenter.api.listing.model_access as listing_model_access
 import ozpcenter.api.library.model_access as library_model_access
 import ozpcenter.api.agency.model_access as agency_model_access
