+++ ozpcenter/api/profile/model_access.py
@@ -7,7 +7,7 @@ from ozpcenter import models
 from django.contrib import auth
 import ozpcenter.model_access as generic_model_access
 
-from plugins_util.plugin_manager import system_has_access_control
+from plugins.plugin_manager import system_has_access_control
 
 # Get an instance of a logger
 logger = logging.getLogger('ozp-center.' + str(__name__))
