+++ plugins/plugin_manager.py
