+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
 DJANGO_SETTINGS_MODULE=ozp.settings
-addopts = -v --cov-report html:cover --cov-report term --cov=ozp --cov=ozpcenter --cov=ozpiwc --cov=plugins --cov=plugins_util
+addopts = -v --cov-report html:cover --cov-report term --cov=ozp --cov=ozpcenter --cov=ozpiwc --cov=plugins
