+++ requirements.txt
@@ -1,3 +1,4 @@
+apipkg==1.4
 appdirs==1.4.3
 autopep8==1.3.2
 backports-abc==0.5
@@ -23,6 +24,7 @@ djangorestframework==3.6.4
 docutils==0.14
 drf-nested-routers==0.90.0
 elasticsearch==5.4.0
+execnet==1.5.0
 factory-boy==2.9.2
 Faker==0.8.4
 flamegraph==0.1
@@ -52,6 +54,8 @@ pyparsing==2.2.0
 pytest==3.2.3
 pytest-cov==2.5.1
 pytest-django==3.1.2
+pytest-forked==0.2
+pytest-xdist==1.20.1
 python-dateutil==2.6.1
 pytz==2017.2
 PyYAML==3.12
